change eventhub management sdk to track2
